requirements-lambda.txt
elevenlabs==2.33.1
pydub
requests
boto3
python-dotenv
